Moved some files, and added _linux.go to them
